[fix]fix the bug of client
