+++ xuluomin/client/srv/show.srv
@@ -1,3 +1,3 @@
-uint8 req
+string name
 ---
 string res
